First experimental version of Complex JSON Handler

+++ back/src/main/scala/com/ebiznext/comet/job/IngestionJob.scala
@@ -1,7 +1,7 @@
 package com.ebiznext.comet.job
 
 import com.ebiznext.comet.config.{DatasetArea, HiveArea, Settings}
-import com.ebiznext.comet.schema.model.{Domain, Metadata, Schema, Write}
+import com.ebiznext.comet.schema.model._
 import org.apache.hadoop.fs.Path
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{DataFrame, Row, SparkSession}
@@ -12,19 +12,23 @@ trait IngestionJob extends SparkJob {
 
   def schema: Schema
 
+  def types: List[Type]
+
   /**
     * Merged metadata
     */
-  val metadata: Metadata = domain.metadata.getOrElse(Metadata()).`import`(schema.metadata.getOrElse(Metadata()))
+  lazy val metadata: Metadata = domain.metadata.getOrElse(Metadata()).`import`(schema.metadata.getOrElse(Metadata()))
 
   /**
     * Dataset loading strategy (JSOn / CSV / ...)
+    *
     * @return Spark Dataframe loaded usingmetadata options
     */
   def loadDataSet(): DataFrame
 
   /**
     * ingestion algorithm
+    *
     * @param dataset
     */
   def ingest(dataset: DataFrame): Unit
@@ -45,38 +49,42 @@ trait IngestionJob extends SparkJob {
     * @param dataset    : dataset to save
     * @param targetPath : absolute path
     * @param writeMode  : Append or overwrite
-    * @param area       : accpeted or rejected area
+    * @param area       : accepted or rejected area
     */
   def saveRows(dataset: DataFrame, targetPath: Path, writeMode: Write, area: HiveArea): Unit = {
-    val count = dataset.count()
-    val saveMode = writeMode.toSaveMode
-    val hiveDB = HiveArea.area(domain.name, area)
-    val tableName = schema.name
-    val fullTableName = s"$hiveDB.$tableName"
-    if (Settings.comet.hive) {
-      logger.info(s"DSV Output $count records to Hive table $hiveDB/$tableName($saveMode) at $targetPath")
-      val dbComment = domain.comment.getOrElse("")
-      session.sql(s"create database if not exists $hiveDB comment '$dbComment'")
-      session.sql(s"use $hiveDB")
-      session.sql(s"drop table if exists $hiveDB.$tableName")
-    }
-
-    val partitionedDF = partitionedDatasetWriter(dataset, metadata.partition.getOrElse(Nil))
-
-    val targetDataset = partitionedDF.mode(saveMode).format("parquet").option("path", targetPath.toString)
-    if (Settings.comet.hive) {
-      targetDataset.saveAsTable(fullTableName)
-      val tableComment = schema.comment.getOrElse("")
-      session.sql(s"ALTER TABLE $fullTableName SET TBLPROPERTIES ('comment' = '$tableComment')")
-      if (Settings.comet.analyze) {
-        val allCols = session.table(fullTableName).columns.mkString(",")
-        val analyzeTable = s"ANALYZE TABLE $fullTableName COMPUTE STATISTICS FOR COLUMNS $allCols"
-        if (session.version.substring(0, 3).toDouble >= 2.4)
-          session.sql(analyzeTable)
+    if (dataset.columns.size > 0) {
+      val count = dataset.count()
+      val saveMode = writeMode.toSaveMode
+      val hiveDB = HiveArea.area(domain.name, area)
+      val tableName = schema.name
+      val fullTableName = s"$hiveDB.$tableName"
+      if (Settings.comet.hive) {
+        logger.info(s"DSV Output $count records to Hive table $hiveDB/$tableName($saveMode) at $targetPath")
+        val dbComment = domain.comment.getOrElse("")
+        session.sql(s"create database if not exists $hiveDB comment '$dbComment'")
+        session.sql(s"use $hiveDB")
+        session.sql(s"drop table if exists $hiveDB.$tableName")
       }
-    }
-    else {
-      targetDataset.save()
+
+      val partitionedDF = partitionedDatasetWriter(dataset, metadata.partition.getOrElse(Nil))
+
+      val targetDataset = partitionedDF.mode(saveMode).format("parquet").option("path", targetPath.toString)
+      if (Settings.comet.hive) {
+        targetDataset.saveAsTable(fullTableName)
+        val tableComment = schema.comment.getOrElse("")
+        session.sql(s"ALTER TABLE $fullTableName SET TBLPROPERTIES ('comment' = '$tableComment')")
+        if (Settings.comet.analyze) {
+          val allCols = session.table(fullTableName).columns.mkString(",")
+          val analyzeTable = s"ANALYZE TABLE $fullTableName COMPUTE STATISTICS FOR COLUMNS $allCols"
+          if (session.version.substring(0, 3).toDouble >= 2.4)
+            session.sql(analyzeTable)
+        }
+      }
+      else {
+        targetDataset.save()
+      }
+    } else {
+      logger.warn("Empty dataset with no columns won't be saved")
     }
   }
 
@@ -87,9 +95,14 @@ trait IngestionJob extends SparkJob {
     * @return : Spark Session used for the job
     */
   def run(args: Array[String]): SparkSession = {
-    schema.presql.getOrElse(Nil).foreach(session.sql)
-    ingest(loadDataSet())
-    schema.postsql.getOrElse(Nil).foreach(session.sql)
+    domain.checkValidity(types) match {
+      case Left(errors) =>
+        errors.foreach(println)
+      case Right(true) =>
+        schema.presql.getOrElse(Nil).foreach(session.sql)
+        ingest(loadDataSet())
+        schema.postsql.getOrElse(Nil).foreach(session.sql)
+    }
     session
   }
